fix creating log files

+++ youtube/Stream.py
@@ -1,9 +1,11 @@
 import threading
 import ffmpeg
+import uuid
 import time
 import re
 import datetime
 import os
+import logging
 import psutil
 
 from tinydb import TinyDB, Query, where
@@ -12,6 +14,8 @@ from tinydb import TinyDB, Query, where
 class Stream:
     STATUS_ACTIVE = 'active'
     STATUS_INACTIVE = 'inactive'
+    STATUS_MANUAL_STOPPED = 'manual_stopped'
+    STATUS_SUCCESS_STOPPED = 'success_stopped'
     TBL_NAME = 'process'
     DB_FILE_PATH = './db/db.json'
 
@@ -29,11 +33,11 @@ class Stream:
         if is_active and psutil.pid_exists(stream_info['pid']):
             return True
 
-        process = self.start_stream(self.input_file, self.output + self.key, self.loop)
-        now = datetime.datetime.now()
-        date_time = now.strftime("%Y%m%d-%H%M%S")
-        log_file = 'ffmpeg-' + date_time + '.log'
+        print("Stream id" + self.key + " has been started")
+        process, log_file = self.start_stream(self.input_file, self.output + self.key, self.loop)
+
         timecode = self.current_timecode(log_file)
+
         db = self.db.table(self.TBL_NAME)
         db.remove((where('key') == self.key))
         db.insert({
@@ -50,20 +54,35 @@ class Stream:
                 re=None,
                 hwaccel='cuvid',
                 # vcodec='h264_cuvid',
-                report=None,
+                # report=None,
                 loglevel='40',
                 stream_loop=loop,
-                ss=timecode
+                ss=timecode,
             ).output(
                 output,
                 vcodec='h264_nvenc',
                 r='30',
                 b="2500k",
                 format='flv'
-            ).run_async(quiet=False)
+            ).run_async(quiet=True, pipe_stdout=True, pipe_stderr=True)
         )
 
-        return process
+        log_file = uuid.uuid4().hex + ".log"
+
+        logger = logging.getLogger(__name__)
+        logger.setLevel(logging.INFO)
+
+        # create a file handler
+        handler = logging.FileHandler(log_file)
+        handler.setLevel(logging.INFO)
+        logger.addHandler(handler)
+
+        out, err = process.communicate()
+
+        logger.info(err.decode())
+        logger.info(out.decode())
+
+        return process, log_file
 
     @staticmethod
     def current_timecode(log_file):
@@ -74,14 +93,19 @@ class Stream:
             text = file.read()
             file.close()
             matches = re.findall(r"time=\w+.\w+.\w+.\w+", text)
-            timecode = matches[-1][5:13]
+
+            if matches:
+                timecode = matches[-1][5:13]
 
         return timecode
 
     @staticmethod
     def is_crush(log_file):
         error_messages = [
-            'libcuda.so.1'
+            'libcuda.so.1',
+            'av_interleaved_write_frame',
+            'Failed creating CUDA',
+            'Invalid data found when processing input',
         ]
         if os.path.isfile(log_file):
             file = open(log_file, 'r')
@@ -91,32 +115,47 @@ class Stream:
                 error = re.search(val, text)
                 if error:
                     return True
-                return False
         return False
 
     def stream_listener(self, process, log_file):
 
         while True:
             time.sleep(self.sleep)
-
-            print(process.pid)
             if process.poll() is not None:
-                if self.is_crush(log_file):
-                    os.remove(log_file)
+                strm = self.get_stream_by_key(self.key)
+                stream = strm[0]
+
+                if self.is_crush(log_file) and stream['status'] != self.STATUS_MANUAL_STOPPED:
+
+                    print("=============================================")
+                    print('завершился не корректно')
+                    print("log file: " + log_file)
                     timecode = self.current_timecode(log_file)
-                    process = self.start_stream(self.input_file, self.output + self.key, self.loop, timecode)
+
+                    print("TIMECODE:" + timecode)
+                    os.remove(log_file)
+
+                    print('Запускаем')
+                    process, log_file = self.start_stream(self.input_file, self.output + self.key, self.loop, timecode)
                     tbl = self.db.table(self.TBL_NAME)
-                    now = datetime.datetime.now()
-                    date_time = now.strftime("%Y%m%d-%H%M%S")
-                    log_file = 'ffmpeg-' + date_time + '.log'
+
                     tbl.update(
                         {'status': self.STATUS_ACTIVE, 'to_start': timecode, 'pid': process.pid, 'log': log_file},
                         where('key') == self.key
                     )
+                    print("=============================================")
                 else:
+                    print("=============================================")
+                    print('завершился корректно')
                     tbl = self.db.table(self.TBL_NAME)
+                    tbl.update(
+                        {'status': self.STATUS_SUCCESS_STOPPED},
+                        where('key') == self.key
+                    )
+                    # time.sleep(30)
                     tbl.remove((Query().pid == process.pid))
                     os.remove(log_file)
+                    print("=============================================")
                     break
 
     def is_active_stream(self, key):
@@ -148,12 +187,14 @@ class Stream:
 
     def stop(self, key):
         strm = self.get_stream_by_key(key)
-        stream = strm[0]
-        os.system("kill -s KILL " + str(stream['pid']))
+        if strm:
+            stream = strm[0]
 
-        # timecode = self.current_timecode(stream['log'])
-        # tbl = self.db.table(self.TBL_NAME)
-        # tbl.update({'status': self.STATUS_INACTIVE, 'to_start': timecode}, where('key') == self.key)
+            os.system("kill -s KILL " + str(stream['pid']))
+
+            timecode = self.current_timecode(stream['log'])
+            tbl = self.db.table(self.TBL_NAME)
+            tbl.update({'status': self.STATUS_MANUAL_STOPPED, 'to_start': timecode}, where('key') == self.key)
 
         return True
